feat(encoder): add tuning and H.264/H.265 switching (#27)
* feat(encoder): edgeconf 채널별 vpuenc_h264 튜닝 키를 읽는다

edgeconf(VHL_CAM.i2cN.chX)에서 설정 가능한 인코더 값이 bps 뿐이었다.
gop 은 CamConfig 필드와 소비 코드는 있는데 JSON 리더가 없어 DEFAULT_GOP_SIZE
에 고정돼 있었고, profile/quant/qp-min/qp-max 는 아예 세팅하지 않아 플러그인
기본값으로 동작했다.

- CamConfig 에 profile/quant/qp_min/qp_max 추가 (bps·gop 과 같은 [rec, rtsp] 쌍)
- parser.cpp json_parser 에 5개 키 리더 추가. gop 리더는 이번에 처음 생긴다
- check_arg() 에 범위 검증(범위 밖이면 로그 + 컴파일 기본값), single-enc 시
  rtsp 슬롯을 rec 으로 정렬, qp_min > qp_max 거부, 채널별 요약 로그
- encoderBin / recordBin / rtspServerBin 세 인코더 경로 모두 값 적용
- enc_set_optional_int(): profile 은 i.MX8MP, qp-min/qp-max 는 i.MX8MP/i.MX8MM
  에서만 플러그인이 install 하므로 property 존재를 확인하고 세팅한다. 없으면
  GLib 경고 대신 "encoder has no '<prop>' property" 로 남기고 건너뛴다

기본값은 전부 종전 실효값과 같다 — 키를 추가하는 것만으로 동작이 바뀌지 않는다.
gop 은 DEFAULT_GOP_SIZE(15) 유지. 이 값이 종전에 4채널이 실제로 쓰던 gop-size 다
(플러그인 자체 기본값 30 이 아니다 — gstApp 이 항상 덮어써 왔다).

함께 고친 기존 결함 — GOptionEntry 의 --grec* 이름/바인딩 불일치:
"grec3" 가 cam[2] 에, "grec0" 가 두 번째로 등록되어 cam[3] 에 묶여 있었다.
결과적으로 --grec2 는 존재하지 않았고 --grec0 은 중복 long name 이었다.
grec2(→cam[2]) / grec3(→cam[3]) 로 정정. --grtsp* 는 원래 정상이다.


* feat(encoder): gop 0 을 fps 연동으로, bitrate 0(VBR) 을 허용한다

앞 커밋에서 gop/bitrate 의 경계값 처리가 어설펐다. VPU 소스를 따라가 확인한
내용을 반영한다.

gop
- DEFAULT_GOP_SIZE 를 0 으로 두고, 0 을 "1초 간격" 센티넬로 정의한다.
  check_arg() 가 해당 스트림의 fps(arg.fps[STREAM_*][ch])로 치환하므로 fps 를
  바꾸면 키프레임 간격이 따라간다. 기본 fps 15 기준 15 로, 종전과 동일하다.
- 0 은 인코더에 넘기지 않는다. gstvpuenc.c 의 주기 I프레임 판정
  (gop_size && gop_count % gop_size)이 0 에서 항상 거짓이 되어 첫 프레임 외에는
  IDR 이 강제되지 않고, wrapper 로 넘어가는 idr_interval=0 의 동작은
  VC8000E 라이브러리가 바이너리라 확인할 수 없다.
- MAX_GOP_SIZE 32767 -> 300. vpu_wrapper_hantro_VCencoder.c:2294 의
  ENC_MAX_GOP_SIZE 가 300 이고 초과분을 조용히 자른다. 상한을 맞춰 gstApp
  로그에서 잘림이 드러나게 한다.

bitrate
- bps 0 을 BITRATE_AUTO 로 받아들인다. vpuenc_h264 는 bitrate 0 을
  "automatic" 으로 문서화하고(자기 기본값도 0), wrapper 의 AdjustBitrate() 가
  해상도·fps 로 목표 레이트를 산출한다. 종전 check_arg() 는 100..20000 밖이라며
  0 을 4096 으로 되돌려 이 모드를 edgeconf 로 시험할 수 없었다.
- 0 이면 기동 시 VBR 임을 로그로 남긴다. 0 이 아닌 값은 종전대로 범위 검증.

채널별 요약 로그를 클램프·센티넬 치환·single-enc 정렬이 모두 끝난 뒤로 옮겼다.
이제 이 라인이 인코더에 실제로 들어가는 값이다.

--grec*/--grtsp* 도움말 8개도 "0=fps(1s), default(0)" 로 갱신.


* fix(encoder): 런타임 set bps/set gop 검증을 edgeconf 경로와 통일한다

같은 값을 edgeconf 로 넣을 때와 런타임 명령으로 넣을 때 판정이 달랐다.

set bps
- 종전 `key < 0 || key > 9999`. 상한이 MAX_BITRATE_KBPS(20000) 와 어긋나 edgeconf
  로는 되는 15000 이 런타임에서는 거부됐다.
- check_arg() 와 같은 규칙으로 통일: BITRATE_AUTO(0) 은 통과, 그 외는
  MIN_BITRATE_KBPS..MAX_BITRATE_KBPS.
- 부수효과: 상한 9999 -> 20000, 하한이 생겨 1..99 는 거부된다(edgeconf 와 동일).

set gop
- 종전 `key < 0 || key > 100`. 상한이 MAX_GOP_SIZE(300) 와 어긋났고, 더 중요하게
  key=0 이 검증을 통과해 g_object_set(enc, "gop-size", 0) 이 그대로 실행됐다.
  앞 커밋에서 0 을 "fps 연동" 센티넬로 정의하고 edgeconf 경로에서만 치환했기
  때문에, 런타임 경로는 같은 0 을 전혀 다른 의미로 처리하고 있었다.
  0 은 인코더에 도달하면 안 되는 값이다(플러그인의 주기 I프레임 판정이 죽는다).
- MIN_GOP_SIZE..MAX_GOP_SIZE 로 검증하고, 적용 직전 enc_gop_resolve() 로
  채널별 fps 치환. fps 는 채널마다 다를 수 있으므로 루프 안에서 채널별로 푼다.


* fix(encoder): bps 범위 검사를 없앤다 — 0 이상이면 통과

앞 커밋에서 런타임 경로를 edgeconf 경로에 맞추면서 하한 100 이 런타임에도
생겼는데, 애초에 이 하한·상한 자체가 근거가 없었다.

wrapper 가 양 끝을 이미 처리한다 (vpu_wrapper_hantro_VCencoder.c AdjustBitrate):
- VPU_ENC_MIN_BITRATE(10000 bps = 10 kbps) 미만 -> 해상도·fps 로 자동 산출.
  즉 0 과 같은 결과가 되므로 하한으로 막을 이유가 없다.
- VPU_ENC_MAX_BITRATE(60000000 bps = 60000 kbps) 초과 -> 클램프.
  gstApp 의 20000 상한은 wrapper 보다 좁아서 멀쩡한 값을 거부하고 있었다.

MIN_BITRATE_KBPS / MAX_BITRATE_KBPS 를 제거하고 네 지점(check_arg rec/rtsp,
set bps rec/rtsp) 모두 음수만 거부한다. BITRATE_AUTO 는 VBR 안내 로그에만 쓴다.


* fix(encoder): setBitrate 파라미터를 gint 로 넓혀 절단을 막는다

Codex 리뷰 지적(P2). setBitrate 의 파라미터가 guint16 인데 호출부는
charArrayToInt() 가 돌려준 gint 를 넘긴다. 종전에는 `set bps` 검증의
`key > 9999` 가 우연히 이 절단을 막고 있었는데, 앞 커밋에서 상한을 없애면서
65535 초과 값이 그대로 guint16 으로 들어가게 됐다.

  set bps rec 0 70000  ->  70000 & 0xFFFF = 4464

"0 이상이면 통과" 라는 의도와 정반대로, 큰 값이 더 낮은 비트레이트로
프로그래밍된다. 세 Bin 의 선언·정의를 gint 로 넓혀 값이 그대로 인코더까지
도달하게 하고, 상한은 wrapper 의 VPU_ENC_MAX_BITRATE(60000 kbps) 클램프에
맡긴다.

setGop 은 guint16 을 유지한다 — MAX_GOP_SIZE 가 300 이라 절단이 불가능하다.


* docs(encoder): quant 만 가드 없이 세팅하는 이유를 주석으로 남긴다

리뷰에서 두 번 나온 질문 — profile/qp-min/qp-max 는 enc_set_optional_int 로
존재 확인 후 세팅하는데 quant 만 g_object_set 을 그대로 쓰는 이유.

gstvpuenc.c 확인 결과 install 조건이 다르다:
- quant                 : in_plugin->std 분기(VPU_V_AVC/MPEG4/H263/VP8/HEVC).
                          코덱 기준이라 MJPEG 외 모든 코덱에 install 된다.
                          gstApp 은 vpuenc_h264(=VPU_V_AVC) 만 쓰므로 항상 존재.
- profile / qp-min/max  : IS_IMX8MP() / IS_IMX8MM() 런타임 SoC 검출에 걸려 있어
                          보드에 따라 install 여부가 갈린다.

즉 quant 는 가드가 불필요한 것이 맞다. 코드 동작 변경은 없고 주석만 추가한다.
encoderBin.cpp 의 기존 주석도 정정했다 — qp-min/qp-max 는 i.MX8MP 전용이
아니라 i.MX8MM 에도 install 된다.


* fix(encoder): 런타임 set bps/set gop 이 cmdArg 를 갱신하도록 한다

리뷰 3인(Claude/Gemini/OpenCode)이 독립적으로 짚은 지적. 런타임 set 명령이
인코더에는 값을 반영하는데 cmdArg 는 그대로 둬서, 이후 cmdArg 를 참조하는
코드가 옛 값을 보게 된다.

조사해 보니 gop 뿐 아니라 bps 도 마찬가지였다. Gemini 는 "bps 는 항상 갱신
한다"고 했지만 실제로는 set bps 의 어느 경로에서도 cmdArg 를 건드리지 않았다.
단일 갱신 지점이던 곳은 set gop rec 의 single_enc 분기 하나뿐이다.

갱신을 추가한 지점:
- set bps rec, single_enc : bps[REC] + bps[RTSP]
  (인코더 하나가 양쪽을 먹이므로 check_arg() 의 미러링과 동일하게 맞춘다)
- set bps rec, dual_enc   : bps[REC]
- set bps rtsp            : bps[RTSP]
- set gop rec, dual_enc   : gop[REC]
- set gop rtsp            : gop[RTSP]

함께 정정 — encoderBin.cpp 주석이 "gop-size stays at DEFAULT_GOP_SIZE (15)"
라고 되어 있었는데, DEFAULT_GOP_SIZE 는 이제 follow-fps 센티넬(0) 이고 15 는
check_arg() 가 fps 로 치환한 결과다. 매크로 리터럴이 아니다.


* feat(encoder): add h264 h265 switching

* fix(encoder): use codec-specific profile defaults

---------

+++ encoderBin.h
@@ -44,7 +44,7 @@ public :
     ~EncoderBin();
 	gboolean init(guint8 ch) ;
     GstPad* getBinSinkPad();
-    void setBitrate(guint16 data);
+    void setBitrate(gint data);
     void getBitrate();
     void setFps(guint16 data);
     void getFps();
